[keepalive] Ensure that display blanking prevention continues after display off. Fixes MER#1422
Blank prevention sessions are started/stopped blindly based on the
state of preventBlanking property. Since mce will ignore all blank prevent
requests made when display is off and/or lockscreen is active, this can
lead to blank prevention renews happening too late and display blanking
even though the preventBlanking is set to true.

Track display and tklock states in addition to the property value. Start
blank prevention session if/when the device enters a state where it is
both requested and allowed.